requirements.txt
Flask==2.3.3
scikit-learn==1.3.0
xgboost==1.7.6
pandas==2.0.3
numpy==1.24.3
tldextract==5.1.1
requests==2.31.0
python-whois==0.8.0
Flask-WTF==1.1.1
WTForms==3.0.1
joblib==1.3.2
urllib3==2.0.4 